+++ test/builders/response-spec.js
@@ -4,6 +4,7 @@ const should = require('should')
 const nassert = require('n-assert')
 const HttpZConsts = require('../../src/consts')
 const HttpZError = require('../../src/error')
+const BaseBuilder = require('../../src/builders/base')
 const ResponseBuilder = require('../../src/builders/response')
 
 describe('builders / response', () => {
@@ -25,7 +26,7 @@ describe('builders / response', () => {
       ResponseBuilder.prototype.build.restore()
     })
 
-    it('should create an instance of ResponseBuilder and call instance.build', () => {
+    it('should create instance of ResponseBuilder and call instance.build', () => {
       let model = {}
       let expected = 'ok'
 
@@ -97,6 +98,72 @@ describe('builders / response', () => {
     })
   })
 
+  describe('_generateHeaderRows', () => {
+    beforeEach(() => {
+      sinon.stub(BaseBuilder.prototype, '_generateHeaderRows')
+    })
+
+    afterEach(() => {
+      BaseBuilder.prototype._generateHeaderRows.restore()
+    })
+
+    it('should throw error when instance.headers is not array', () => {
+      let builder = new ResponseBuilder({ headers: 'incorrect headers' })
+
+      should(builder._generateHeaderRows.bind(builder)).throw(HttpZError, {
+        message: 'headers must be an array'
+      })
+
+      nassert.assertFn({ inst: BaseBuilder.prototype, fnName: '_generateHeaderRows' })
+    })
+
+    it('should remove set-cookie headers and call parent method', () => {
+      let builder = new ResponseBuilder({
+        headers: [
+          {
+            name: 'set-cookie',
+            values: 'some cookie1'
+          },
+          {
+            name: 'connection',
+            values: []
+          },
+          {
+            name: 'accept',
+            values: [
+              { value: '*/*' }
+            ]
+          },
+          {
+            name: 'set-cookie',
+            values: 'some cookie2'
+          }
+        ]
+      })
+      let expected = 'ok'
+      let expectedHeaders = [
+        {
+          name: 'connection',
+          values: []
+        },
+        {
+          name: 'accept',
+          values: [
+            { value: '*/*' }
+          ]
+        }
+      ]
+
+      BaseBuilder.prototype._generateHeaderRows.returns('ok')
+
+      let actual = builder._generateHeaderRows()
+      should(actual).eql(expected)
+      should(builder.headers).eql(expectedHeaders)
+
+      nassert.assertFn({ inst: BaseBuilder.prototype, fnName: '_generateHeaderRows', expectedArgs: '_without-args_' })
+    })
+  })
+
   describe('_generateCookieRows', () => {
     function getDefaultCookes() {
       return [
@@ -105,14 +172,6 @@ describe('builders / response', () => {
         { name: 'username', value: 'smith', params: ['Expires=Wed, 21 Oct 2015 07:28:00 GMT', 'Secure', 'HttpOnly'] }
       ]
     }
-
-    it('should return empty string when instance.cookies is undefined', () => {
-      let builder = getBuilderInstance({ cookies: undefined })
-
-      let expected = ''
-      let actual = builder._generateCookieRows()
-      should(actual).eql(expected)
-    })
 
     it('should throw error when instance.cookies is not array', () => {
       let builder = getBuilderInstance({ cookies: 'wrong cookies' })
@@ -144,6 +203,22 @@ describe('builders / response', () => {
         message: 'cookie params must be an array',
         details: 'cookie index: 1'
       })
+    })
+
+    it('should return empty string when instance.cookies is undefined', () => {
+      let builder = getBuilderInstance({ cookies: undefined })
+
+      let expected = ''
+      let actual = builder._generateCookieRows()
+      should(actual).eql(expected)
+    })
+
+    it('should return empty string when instance.cookies is empty array', () => {
+      let builder = getBuilderInstance({ cookies: [] })
+
+      let expected = ''
+      let actual = builder._generateCookieRows()
+      should(actual).eql(expected)
     })
 
     it('should build cookie rows when instance.cookies are valid', () => {
@@ -194,7 +269,7 @@ describe('builders / response', () => {
         ]
       }
 
-      let plainResponse = [
+      let rawResponse = [
         'HTTP/1.1 201 Created',
         'Connection: ',
         'Cache-Control: no-cache',
@@ -206,7 +281,7 @@ describe('builders / response', () => {
 
       let builder = getBuilderInstance(responseModel)
       let actual = builder.build()
-      should(actual).eql(plainResponse)
+      should(actual).eql(rawResponse)
     })
 
     it('should build response with cookies, but without body', () => {
@@ -247,7 +322,7 @@ describe('builders / response', () => {
         ]
       }
 
-      let plainResponse = [
+      let rawResponse = [
         'HTTP/1.1 201 Created',
         'Connection: ',
         'Cache-Control: no-cache',
@@ -263,7 +338,7 @@ describe('builders / response', () => {
 
       let builder = getBuilderInstance(responseModel)
       let actual = builder.build()
-      should(actual).eql(plainResponse)
+      should(actual).eql(rawResponse)
     })
 
     it('should build response with body of contentType=text/plain', () => {
@@ -310,7 +385,7 @@ describe('builders / response', () => {
         }
       }
 
-      let plainResponse = [
+      let rawResponse = [
         'HTTP/1.1 200 Ok',
         'Connection: keep-alive',
         'Cache-Control: no-cache',
@@ -323,152 +398,7 @@ describe('builders / response', () => {
 
       let builder = getBuilderInstance(responseModel)
       let actual = builder.build()
-      should(actual).eql(plainResponse)
-    })
-
-    it('should build response with body of contentType=multipart/alternative (inline)', () => {
-      let responseModel = {
-        protocolVersion: 'HTTP/1.1',
-        statusCode: 200,
-        statusMessage: 'Ok',
-        headers: [
-          {
-            name: 'Connection',
-            values: [
-              { value: 'keep-alive' }
-            ]
-          },
-          {
-            name: 'Cache-Control',
-            values: [
-              { value: 'no-cache' }
-            ]
-          },
-          {
-            name: 'Content-Encoding',
-            values: [
-              { value: 'gzip' },
-              { value: 'deflate' }
-            ]
-          },
-          {
-            name: 'Content-Length',
-            values: [
-              { value: '301' }
-            ]
-          },
-          {
-            name: 'Content-Type',
-            values: [
-              { value: 'multipart/alternative', params: 'boundary="111362-53119209"' }
-            ]
-          }
-        ],
-        body: {
-          contentType: 'multipart/alternative',
-          boundary: '111362-53119209',
-          params: [
-            {
-              type: 'inline',
-              value: '<base64-data>'
-            }
-          ]
-        },
-        headersSize: 243,
-        bodySize: 84
-      }
-      let plainResponse = [
-        'HTTP/1.1 200 Ok',
-        'Connection: keep-alive',
-        'Cache-Control: no-cache',
-        'Content-Encoding: gzip, deflate',
-        'Content-Length: 301',
-        'Content-Type: multipart/alternative;boundary="111362-53119209"',
-        '',
-        '--111362-53119209',
-        'Content-Disposition: inline',
-        '',
-        '<base64-data>',
-        '--111362-53119209--'
-      ].join(HttpZConsts.EOL)
-
-      let builder = getBuilderInstance(responseModel)
-      let actual = builder.build()
-      should(actual).eql(plainResponse)
-    })
-
-    it('should build response with body of contentType=multipart/mixed (attachment)', () => {
-      let responseModel = {
-        protocolVersion: 'HTTP/1.1',
-        statusCode: 200,
-        statusMessage: 'Ok',
-        headers: [
-          {
-            name: 'Connection',
-            values: [
-              { value: 'keep-alive' }
-            ]
-          },
-          {
-            name: 'Cache-Control',
-            values: [
-              { value: 'no-cache' }
-            ]
-          },
-          {
-            name: 'Content-Encoding',
-            values: [
-              { value: 'gzip' },
-              { value: 'deflate' }
-            ]
-          },
-          {
-            name: 'Content-Length',
-            values: [
-              { value: '301' }
-            ]
-          },
-          {
-            name: 'Content-Type',
-            values: [
-              { value: 'multipart/mixed', params: 'boundary="11136253119209"' }
-            ]
-          }
-        ],
-        body: {
-          contentType: 'multipart/mixed',
-          boundary: '11136253119209',
-          params: [
-            {
-              type: 'attachment',
-              contentType: 'application/octet-stream',
-              fileName: 'photo1.jpg',
-              value: '<binary-data>'
-            }
-          ]
-        },
-        headersSize: 236,
-        bodySize: 149
-      }
-      let plainResponse = [
-        'HTTP/1.1 200 Ok',
-        'Connection: keep-alive',
-        'Cache-Control: no-cache',
-        'Content-Encoding: gzip, deflate',
-        'Content-Length: 301',
-        'Content-Type: multipart/mixed;boundary="11136253119209"',
-        '',
-        '--11136253119209',
-        'Content-Disposition: attachment; filename="photo1.jpg"',
-        'Content-Type: application/octet-stream',
-        '',
-        '<binary-data>',
-        '--11136253119209--'
-      ].join(HttpZConsts.EOL)
-
-      let builder = getBuilderInstance(responseModel)
-      let actual = builder.build()
-      should(actual).eql(plainResponse)
+      should(actual).eql(rawResponse)
     })
   })
 })